requirements.txt
Django==1.7.1
pytz==2014.7
unicodecsv==0.9.4

django-cms==3.0.7
djangocms-text-ckeditor==2.4.1

# Optional, recommended packages
Pillow>=2
django-filer==0.9.8
cmsplugin-filer==0.10.1
django-reversion

# Django CMS Blog
djangocms-blog

# Event list on the homepage (subscribe_tags)
icalendar==3.8.4
requests==2.5.1

# Django compressor
django-compressor==1.4
cssmin==0.2.0

# Caching
python-memcached==1.53

# Registration, so people can create themselves accounts
django-registration-redux

# Markdown, to format article proposals
Markdown>=2

# Application publish
django-admin-sortable2
djangorestframework
django-filter
django-cors-headers==1.1.0
